requirements.txt
asgiref==3.5.0
Django==3.0.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
pymongo==3.12.2
pytz==2022.1
sqlparse==0.2.4
